requirements.txt
asgiref==3.4.1
Django==3.2.5
django-adminlte3==0.1.6
django-login-required-middleware==0.6.1
pytz==2021.1
sqlparse==0.4.1
